Update: Adding Wagtail object type compatible with PageInterface
