+++ baseapp_wagtail/base/graphql/object_types.py
@@ -1,9 +1,25 @@
 import graphene
+from graphene import relay
 
-from baseapp_wagtail.base.graphql.interfaces import URLPathTargetWagtailInterface
+from baseapp_pages.graphql import PageInterface
+from baseapp_wagtail.base.graphql.interfaces import WagtailPageInterface
 
 
-class WagtailPageObjectType(graphene.ObjectType):
+class WagtailURLPathObjectType(graphene.ObjectType):
+    data = graphene.Field(WagtailPageInterface)
+
     class Meta:
-        interfaces = (URLPathTargetWagtailInterface,)
+        interfaces = (
+            relay.Node,
+            PageInterface,
+        )
         name = "WagtailPage"
+
+    @classmethod
+    def resolve_type(cls, instance, info):
+        return cls
+
+    def resolve_data(self, info):
+        return self
+
+    # TODO: (BA-2636) Solve the metadata interface.
